// web/src/gen/model/tagExportPackageDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { ListOfTagExportItemDto } from "./listOfTagExportItemDto";

export interface TagExportPackageDto {
  formatVersion: string;
  exportedAt: string;
  totalTags: number;
  tags: ListOfTagExportItemDto;
}