// web/src/gen/model/tagExportItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */

export interface TagExportItemDto {
  path: string;
  name: string;
  /** @nullable */
  color: string | null;
  /** @nullable */
  description: string | null;
}