// web/src/gen/model/importTagsCommand.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { TagConflictStrategy } from "./tagConflictStrategy";

export interface ImportTagsCommand {
  /** @minLength 1 */
  projectId: string;
  conflictStrategy?: TagConflictStrategy;
  /** @nullable */
  tags?: unknown[] | null;
  /** @nullable */
  csvContent?: string | null;
}