// web/src/gen/model/importTagsResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation Studio
 * Digital Content Creation Automation Studio Engine & API
 * OpenAPI spec version: 1.0.0
 */
import type { ListOfTagImportError } from "./listOfTagImportError";

export interface ImportTagsResult {
  totalProcessed: number;
  createdCount: number;
  updatedCount: number;
  skippedCount: number;
  errors: ListOfTagImportError;
}